+++ contracts/test/accumulators/CurveLiquidityAccumulatorStub.sol
@@ -25,8 +25,8 @@ contract CurveLiquidityAccumulatorStub is CurveLiquidityAccumulator {
 
     function validateObservation(
         address,
-        uint256,
-        uint256
+        uint112,
+        uint112
     ) internal virtual override returns (bool) {
         return true; // Disable for simplicity
     }
