+++ contracts/test/accumulators/UniswapV3LiquidityAccumulatorStub.sol
@@ -36,8 +36,8 @@ contract UniswapV3LiquidityAccumulatorStub is UniswapV3LiquidityAccumulator {
 
     function validateObservation(
         address,
-        uint256,
-        uint256
+        uint112,
+        uint112
     ) internal virtual override returns (bool) {
         return true; // Disable for simplicity
     }
